pyproject.toml
[tool.poetry]
name = "ai-utilities"
version = "0.1.0"
description = ""
authors = ["Francesc-Sanchez <54701409+Francesc-Sanchez@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langgraph = "^1.0.2"
python-dotenv = "^1.2.1"
langchain-deepseek = "^1.0.0"
langchain-openai = "^1.0.2"
biopython = "^1.86"
beautifulsoup4 = "^4.14.2"
ddgs = "^9.8.0"
langchain-community = "^0.4.1"
google-search-results = "^2.4.2"
arxiv = "==2.3.0"
langchain = "^1.0.5"
wikipedia = "==1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
